ESP32 support

+++ airrohr-firmware/utils.cpp
@@ -21,15 +21,12 @@
  ************************************************************************
  */
 
-#include <WString.h>
-
-#include "./intl.h"
-#include "./utils.h"
-#include "./defines.h"
-#include "./ext_def.h"
+#include "intl.h"
+#include "utils.h"
+#include "defines.h"
+#include "ext_def.h"
 
 #include "ca-root.h"
-
 
 /*****************************************************************
  * aircms.online helper functions                                *
@@ -217,7 +214,7 @@ bool launchUpdateLoader(const String& md5) {
  *****************************************************************/
 String check_display_value(double value, double undef, uint8_t len, uint8_t str_len) {
 	RESERVE_STRING(s, 15);
-	s = (value != undef ? String(value, len) : String("-"));
+	s = (value != undef ? String(value, static_cast<unsigned int>(len)) : String("-"));
 	while (s.length() < str_len) {
 		s = " " + s;
 	}
